requirements.txt
psycopg2
python-dotenv
slack_sdk